Removed all links to javafx.scene in the API
Everything in the javafx.scene package is somehow connected to the view.
As the API should not be in charge of the user interface (the view),
nothing anymore uses this package. By the way, the dependency package.
Instead, the javafx-base has been added as it contains observable values
useful in any API.